Break parsing logic into files Convert transform-parsoid into module
