+++ lambdas/transform-parsoid/go.mod
@@ -1,0 +1,8 @@
+module transform-parsoid
+
+go 1.14
+
+require (
+	github.com/PuerkitoBio/goquery v1.5.1
+	golang.org/x/net v0.0.0-20200625001655-4c5254603344
+)
