+++ lambdas/transform-parsoid/reference.go
@@ -1,0 +1,4 @@
+package main
+
+type Reference struct {
+}
